requirements.txt
pyyaml===6.01
pytest===8.1.1